+++ Changelog.md
@@ -57,6 +57,7 @@ We use yarn to install the frontend dependencies now, so you need to have that i
 * Upgrade to rails 6.1 [#8366](https://github.com/diaspora/diaspora/pull/8366)
 * Update the suggested Ruby version to 2.7. If you run into trouble during the update and you followed our installation guides, run `rvm install 2.7`. [#8366](https://github.com/diaspora/diaspora/pull/8366)
 * Upgrade to bundler 2 [#8366](https://github.com/diaspora/diaspora/pull/8366)
+* Stop checking `/.well-known/host-meta`, check for `/.well-known/nodeinfo` instead [#8377](https://github.com/diaspora/diaspora/pull/8377)
 
 ## Bug fixes
 * Fix that no mails were sent after photo export [#8365](https://github.com/diaspora/diaspora/pull/8365)